+++ src/shinestacker/app/main.py
@@ -1,30 +1,26 @@
 # pylint: disable=C0114, C0115, C0116, C0413, E0611, R0903, E1121, W0201, R0915, R0912
 import sys
-import os
-import logging
 import argparse
 import matplotlib
 import matplotlib.backends.backend_pdf
 matplotlib.use('agg')
 from PySide6.QtWidgets import (QApplication, QMainWindow, QVBoxLayout, QHBoxLayout, QStackedWidget,
                                QMenu, QMessageBox, QDialog, QLabel, QListWidget, QPushButton)
-from PySide6.QtGui import QAction, QIcon, QGuiApplication
-from PySide6.QtCore import Qt, QEvent, QTimer, Signal
+from PySide6.QtGui import QAction, QGuiApplication
+from PySide6.QtCore import QEvent, QTimer, Signal
 from shinestacker.config.config import config
 config.init(DISABLE_TQDM=True, COMBINED_APP=True, DONT_USE_NATIVE_MENU=True)
 from shinestacker.config.constants import constants
-from shinestacker.config.settings import StdPathFile
-from shinestacker.core.logging import setup_logging
 from shinestacker.gui.main_window import MainWindow
 from shinestacker.retouch.image_editor_ui import ImageEditorUI
-from shinestacker.app.gui_utils import (
-    disable_macos_special_menu_items, fill_app_menu, set_css_style)
+from shinestacker.app.gui_utils import fill_app_menu
 from shinestacker.app.help_menu import add_help_action
 from shinestacker.app.open_frames import open_frames
 from shinestacker.app.args_parser_opts import (
     add_project_arguments, add_retouch_arguments, extract_positional_filename,
     setup_filename_argument, process_filename_argument
 )
+from shinestacker.app.gui_utils import make_app
 
 
 class SelectionDialog(QDialog):
@@ -230,18 +226,11 @@ open retouch window at startup instead of project windows.
     if filename and path:
         print("can't specify both arguments --filename and --path", file=sys.stderr)
         sys.exit(1)
-    setup_logging(console_level=logging.DEBUG, file_level=logging.DEBUG, disable_console=True,
-                  log_file=StdPathFile('shinestacker.log').get_file_path())
-    app = Application(sys.argv)
-    if config.DONT_USE_NATIVE_MENU:
-        app.setAttribute(Qt.AA_DontUseNativeMenuBar)
-    else:
-        disable_macos_special_menu_items()
-    icon_path = f"{os.path.dirname(__file__)}/../gui/ico/shinestacker.png"
-    app.setWindowIcon(QIcon(icon_path))
+
+    app = make_app(Application)
     main_app = MainApp()
     app.main_app = main_app
-    set_css_style(app)
+
     main_app.show()
     main_app.activateWindow()
     if args['expert']: