+++ src/shinestacker/app/retouch.py
@@ -1,25 +1,20 @@
 # pylint: disable=C0114, C0115, C0116, C0413, E0611, R0903, E1121, W0201
-import os
 import sys
-import logging
 import argparse
 from PySide6.QtWidgets import QApplication, QMenu
-from PySide6.QtGui import QIcon
-from PySide6.QtCore import Qt, QEvent
+from PySide6.QtCore import QEvent
 from shinestacker.config.config import config
 config.init(DISABLE_TQDM=True, DONT_USE_NATIVE_MENU=True)
 from shinestacker.config.constants import constants
-from shinestacker.config.settings import StdPathFile
-from shinestacker.core.logging import setup_logging
 from shinestacker.retouch.image_editor_ui import ImageEditorUI
-from shinestacker.app.gui_utils import (
-    disable_macos_special_menu_items, fill_app_menu, set_css_style)
+from shinestacker.app.gui_utils import fill_app_menu
 from shinestacker.app.help_menu import add_help_action
 from shinestacker.app.open_frames import open_frames
 from shinestacker.app.args_parser_opts import (
     add_retouch_arguments, extract_positional_filename,
     setup_filename_argument, process_filename_argument
 )
+from shinestacker.app.gui_utils import make_app
 
 
 class RetouchApp(ImageEditorUI):
@@ -45,32 +40,20 @@ class Application(QApplication):
 
 
 def main():
-    positional_filename, filtered_argv = extract_positional_filename()
-    original_argv = sys.argv
-    sys.argv = filtered_argv
+    positional_filename, filtered_args = extract_positional_filename()
     parser = argparse.ArgumentParser(
         prog=f'{constants.APP_STRING.lower()}-retouch',
         description='Final retouch focus stack image from individual frames.',
         epilog=f'This app is part of the {constants.APP_STRING} package.')
     setup_filename_argument(parser, use_const=True)
     add_retouch_arguments(parser)
-    args = vars(parser.parse_args())
-    sys.argv = original_argv
+    args = vars(parser.parse_args(filtered_args))
     filename = process_filename_argument(args, positional_filename)
     path = args['path']
     if filename and path:
         print("can't specify both arguments --filename and --path", file=sys.stderr)
         sys.exit(1)
-    setup_logging(console_level=logging.DEBUG, file_level=logging.DEBUG, disable_console=True,
-                  log_file=StdPathFile('shinestacker.log').get_file_path())
-    app = Application(sys.argv)
-    if config.DONT_USE_NATIVE_MENU:
-        app.setAttribute(Qt.AA_DontUseNativeMenuBar)
-    else:
-        disable_macos_special_menu_items()
-    icon_path = f"{os.path.dirname(__file__)}/../gui/ico/shinestacker.png"
-    app.setWindowIcon(QIcon(icon_path))
-    set_css_style(app)
+    app = make_app(Application)
     editor = RetouchApp()
     app.editor = editor
     editor.show()