+++ src/shinestacker/retouch/denoise_filter.py
@@ -6,7 +6,7 @@ from .. algorithms.denoise import denoise
 class DenoiseFilter(OneSliderBaseFilter):
     def __init__(self, name, parent, image_viewer, layer_collection, undo_manager):
         super().__init__(name, parent, image_viewer, layer_collection, undo_manager,
-                         10.0, 2.5, "Denoise",
+                         0.0, 10.0, 2.5, "Denoise",
                          allow_partial_preview=True, preview_at_startup=False)
 
     def apply(self, image, strength):